sooo close to working:
the only problem is that the command key is still applying to the next key press. so if you make sure to let go of right command before pressing the next key for your shortcut, then this works totally fine. but if you like to roll through keypresses like me, this won't work because instead of leader r you get leader cmd+r